backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
supabase==2.7.4
pytrends==4.9.2
google-generativeai==0.8.3
google-ads==24.1.0
python-multipart==0.0.9
